// React/app/GerAdv_TS/AnexamentoRegistros/Interfaces/interface.AnexamentoRegistros.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAnexamentoRegistros {
// 202501251
    id: number;
 
	cliente: number,
	guidreg: string,
	codigoreg: number,
	idreg: number,
	data: string,
}

export interface IAnexamentoRegistrosFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAnexamentoRegistrosIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}